ai/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.1
chromadb>=0.4.6
mysql-connector-python>=8.0.33
asyncio>=3.4.3
websockets>=11.0.3